+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Ability for `conduit-translate` to deal with chunks from `conduit-collect`.
+
 ### Fixed
 - Help message for `-chunk_size` in `conduit-collect`. Previously was ambiguous,
   saying `-chunk_size` corresponded to number of files, instead of size of files.